tests: ido: rework test for option -P
The test strips off the first GID (i.e. 0 for root) from the ido
environment to compare the list of group ids given by the bash array
$GROUPS to the one from host environment.

	gportay@archlinux ~/src/iamroot $ echo "${GROUPS[@]}"
	1000 965 966 968 973 974 986 990 991 994 998 150 1001

	gportay@archlinux ~/src/iamroot $ ido -P -s
	archlinux ~/src/iamroot # echo "${GROUPS[@]}"
	0 1000 965 966 968 973 974 986 990 991 994 998 150 1001

This reworks the test for the option -P by removing the strip for the
root GID (ido environment) and prepending its hardcoded value (0) to the
list of the expected group ids (host environment).