tests: ish: set bash to test for $-
The default shell (/bin/sh) set by the different Unix systems is
different.

Here is a bit of history about the shells...

The Unix operating system used the Thompson shell until Version 6 Unix,
and moved to the Bourne shell since Version 7 Unix.

The BSD operating systems uses different shells; the NetBSD and FreeBSD
systems use the Alquimist shell[1][2]; the OpenBSD system uses the Korn
shell[3][4]; All of them ships a C shell (NetBSD and OpenBSD) or the
TENNEX C shell (FreeBSD); the csh can be used for interactive shells,
and the sh remains used for shell scripting.

Note: The /bin/csh is the default root shell on FreeBSD 13 and earlier;
FreeBSD 14 switches to /bin/sh to match the default user shell.

The GNU/Linux operating systems uses the Bourne again shell[5] or the
Debian Almquist shell[6] (Debian derivatives).

Also, Busybox forked the Almquist shell[7] (and hush[8] before that).

Now, step back to the root issue...

The variable $- is neither implemented by the C shell nor by the TENNEX
C shell; it is even considered as an illegal variable name:

	freebox@freebsd:~ % uname -s
	FreeBSD
	freebox@freebsd:~ % getent passwd root
	root::0:0:Charlie &:/root:/bin/csh
	freebox@freebsd:~ % echo "$SHELL"
	/bin/csh
	freebox@freebsd:~ % echo "$-"
	Illegal variable name.
	freebox@freebsd:~ % md5 /bin/sh /bin/ksh /bin/csh /bin/tcsh
	MD5 (/bin/sh) = 0db50b8ffff78a6a43071797bb076933
	md5: /bin/ksh: No such file or directory
	MD5 (/bin/csh) = 06131f58034f164091f13004b557dd4b
	MD5 (/bin/tcsh) = 06131f58034f164091f13004b557dd4b

	arm64$ uname -s
	NetBSD
	arm64$ getent passwd root
	root:*:0:0:Charlie &:/root:/bin/sh
	arm64$ echo "$SHELL"
	/bin/sh
	arm64$ echo "$-"
	EiLlms
	arm64$ md5 /bin/sh /bin/ksh /bin/csh
	MD5 (/bin/sh) = 8a8931a2371a6cd07b3f53774d22a7a1
	MD5 (/bin/ksh) = 42124531c4bdbe984f0429b691d4a53d
	MD5 (/bin/csh) = 894b71a09db7e1dbb047ff4b92bc6db4

	openbsd$ uname -s
	OpenBSD
	openbsd$ getent passwd root
	root:*:0:0:Charlie &:/root:/bin/ksh
	openbsd$ echo "$SHELL"
	/bin/sh
	openbsd$ echo "$-"
	ilms
	openbsd$ md5 /bin/sh /bin/ksh /bin/csh
	MD5 (/bin/sh) = f39a5099deee85812d00bf9934973e21
	MD5 (/bin/ksh) = f39a5099deee85812d00bf9934973e21
	MD5 (/bin/csh) = c836cd714d1f7de70087eb7a49e45191

Although, the output of the variable $- differs through the shells as
they implements different options. Moreover, they enables differents
options by default.

As a consequence, the tests using the content of variable $- **MUST**
ensure to use the shell expecting the output; i.e. bash in this case.

This sets bash to the tests using the content of variable $- as that
variable is not implemented on csh, the default shell used on FreeBSD
systems (actually tcsh), and as that variable outputs different values
on ash and ksh used on NetBSD and OpenBSD respectively.

[1]: https://github.com/NetBSD/src/tree/trunk/bin/sh
[2]: https://github.com/freebsd/freebsd-src/tree/main/bin/sh
[3]: https://github.com/openbsd/src/commit/1ea725044790f9d21f338ca09b35a7b732bd2361
[4]: https://github.com/openbsd/src/commit/e13666c96ae124b25f283ffa6aaabb858a4f88a4
[5]: http://git.savannah.gnu.org/cgit/bash.git/tree/
[6]: https://git.kernel.org/pub/scm/utils/dash/dash.git
[7]: https://git.busybox.net/busybox/tree/shell/ash.c
[8]: https://git.busybox.net/busybox/tree/shell/hush.c