Fix "tests: ido: rework test for option -P"
The Github Actions FreeBSD-vm, NetBSD-vm and OpenBSD-vm run the CI as
root.

The test prepends the hardcoded value of root (i.e. 0) to the list of
the expected GROUPS.

That value is already part of the GROUPS since ido does not "switch" to
another user (i.e. it runs as root and it usurps the root identity).

As a consequence, the root user-id is duplicated and raises the error
attached below.

This fixes commit 657c1bea9c5f4a70570b781376770580d462b165.

Fixes:

	Running ido: test option -P preserves group vector 0 0 0 5...
	GROUPS=0 0 5
	ido: test option -P preserves group vector 0 0 0 5: [KO]